create all signal graph and audio frame functionality in headers
the previous implementaiton tried to use template metaprogramming but I ran into weirdness when a node had only one arity. I think that if I try this again I can have a separate inputArity and outputArity to get around the problems I was having. if I know for certain that the input node has an output arity that is compatible with the input arity of the receiving node then I can bypass the bug I ran into.