Convert RtMidi dependency from git submodule to vcpkg package
RtMidi places it's headers in `include/rtmidi` instead of directly in `include`. This causes me to need to import it with a brittle path that enforces the dependency must come from vcpkg. Vcpkg is the tool used for packages moving forward, but I still think this is a bad practice. RtMidi has been buggy for me anyway, so it's ripe for replacing.