Remove the CoefficientStrategySettable class for Filter and EQ nodes because the strategy is going to be fixed to RBJ Biquad coefficients.
The interchangeable strategy approach erroneously assumed that different methods of calculating coefficients for biquad filters all had the same interface. I have since learned that this is not the case. The other filter formulae (butterworth, etc) may be better suited as a component of a more complex algorithm rather than a stand-alone Node.